+++ server/contract/group_test.go
@@ -19,7 +19,7 @@ func TestGroup(t *testing.T) {
 		MaxAttendees: 5,
 		Attendees:    4,
 		HasResponded: true,
-		SearchHints:  "George Hoover",
+		SearchHints:  "George Hoover, Tom Thumb",
 	}
 	group2 := ced.Group{
 		ID:           ced.NewID(),
@@ -27,7 +27,7 @@ func TestGroup(t *testing.T) {
 		MaxAttendees: 2,
 		Attendees:    0,
 		HasResponded: false,
-		SearchHints:  "Elizabeth Hoover",
+		SearchHints:  "Elizabeth Hoover, Tom Thumb",
 	}
 	reqCtx := ced.ReqContext{}
 
@@ -120,7 +120,7 @@ func TestGroup(t *testing.T) {
 			defer setup(t)()
 
 			err := groupContract.Respond(context.Background(), reqCtx, group1.ID, 6)
-			testutils.IsCodeAndError(t, err, ced.EINVALID, "group can have at most 5 attendees")
+			testutils.IsCodeAndError(t, err, ced.EINVALID, "group can have at most 5 attendees, has 6")
 		})
 	})
 
@@ -140,6 +140,188 @@ func TestGroup(t *testing.T) {
 
 			_, err := groupContract.Get(context.Background(), reqCtx, ced.NewID())
 			testutils.IsCode(t, err, ced.ENOTFOUND)
+		})
+	})
+
+	t.Run("find one", func(t *testing.T) {
+
+		t.Run("can find one group", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			res, err := groupContract.FindOne(context.Background(), "Goerge Hoover and family")
+			is.NoErr(err)
+			is.Equal(res, group1)
+		})
+
+		t.Run("handles if nothing found", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			_, err := groupContract.FindOne(context.Background(), "nonsense")
+			is.Equal("find one group: no results", err.Error())
+		})
+
+		t.Run("handles too many results", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			_, err := groupContract.FindOne(context.Background(), "Tom Thumb")
+			is.Equal("find one group: too many results, try narrowing the search", err.Error())
+		})
+	})
+
+	t.Run("update group", func(t *testing.T) {
+
+		t.Run("handles group not found", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{ID: ced.NewID()}
+
+			err := groupContract.Update(context.Background(), update)
+			is.Equal("update group: Not found", err.Error())
+		})
+
+		t.Run("validates name", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:   group1.ID,
+				Name: pointerize(ced.Name("")),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.Equal("update group: Name is required.", err.Error())
+		})
+
+		t.Run("validates attendees update", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:        group1.ID,
+				Attendees: pointerize(uint8(6)),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.Equal("update group: group can have at most 5 attendees, has 6", err.Error())
+		})
+
+		t.Run("validates max attendees update", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:           group1.ID,
+				MaxAttendees: pointerize(uint8(0)),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.Equal("update group: group can have at most 0 attendees, has 4", err.Error())
+		})
+
+		t.Run("validates attendees and max attendees update", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:           group1.ID,
+				Attendees:    pointerize(uint8(2)),
+				MaxAttendees: pointerize(uint8(1)),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.Equal("update group: group can have at most 1 attendees, has 2", err.Error())
+		})
+
+		t.Run("can update name", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:   group1.ID,
+				Name: pointerize(ced.Name("The Muffin Man")),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.NoErr(err)
+
+			res, err := groupRepository.Get(context.Background(), group1.ID)
+			is.NoErr(err)
+
+			is.Equal(ced.Name("The Muffin Man"), res.Name)
+			is.Equal(group1.Attendees, res.Attendees)
+			is.Equal(group1.MaxAttendees, res.MaxAttendees)
+			is.Equal(group1.HasResponded, res.HasResponded)
+			is.Equal(group1.SearchHints, res.SearchHints)
+		})
+
+		t.Run("can update attendees", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:        group2.ID,
+				Attendees: pointerize(uint8(1)),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.NoErr(err)
+
+			res, err := groupRepository.Get(context.Background(), group2.ID)
+			is.NoErr(err)
+
+			is.Equal(group2.Name, res.Name)
+			is.Equal(uint8(1), res.Attendees)
+			is.Equal(group2.MaxAttendees, res.MaxAttendees)
+			is.Equal(true, res.HasResponded)
+			is.Equal(group2.SearchHints, res.SearchHints)
+		})
+
+		t.Run("can update max attendees", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:           group2.ID,
+				MaxAttendees: pointerize(uint8(10)),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.NoErr(err)
+
+			res, err := groupRepository.Get(context.Background(), group2.ID)
+			is.NoErr(err)
+
+			is.Equal(group2.Name, res.Name)
+			is.Equal(group2.Attendees, res.Attendees)
+			is.Equal(uint8(10), res.MaxAttendees)
+			is.Equal(false, res.HasResponded)
+			is.Equal(group2.SearchHints, res.SearchHints)
+		})
+
+		t.Run("can update search hints", func(t *testing.T) {
+			defer setup(t)()
+			is := is.New(t)
+
+			update := ced.GroupUpdate{
+				ID:          group2.ID,
+				SearchHints: pointerize("here is a hint"),
+			}
+
+			err := groupContract.Update(context.Background(), update)
+			is.NoErr(err)
+
+			res, err := groupRepository.Get(context.Background(), group2.ID)
+			is.NoErr(err)
+
+			is.Equal(group2.Name, res.Name)
+			is.Equal(group2.Attendees, res.Attendees)
+			is.Equal(group2.MaxAttendees, res.MaxAttendees)
+			is.Equal(group2.HasResponded, res.HasResponded)
+			is.Equal("here is a hint", res.SearchHints)
 		})
 	})
 
@@ -207,4 +389,8 @@ func TestGroup(t *testing.T) {
 			is.Equal(group2, resGroup2)
 		})
 	})
+}
+
+func pointerize[T any](t T) *T {
+	return &t
 }