+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.8.0
 
 - Simplify UI of RSVP page
+- Add `ced group update` command
 
 Full changelog: [v0.7.0...v0.8.0](https://github.com/bradenrayhorn/ced/compare/v0.7.0...v0.8.0)